+++ internal/entities/prices_response.go
@@ -1,7 +1,7 @@
 package entities
 
 type PriceResponse struct {
-	VehicleType     string `json:"vehicle_type"`
-	ReservationTime string `json:"reservation_time"`
-	Price           int    `json:"price"`
+	VehicleType     string  `json:"vehicle_type"`
+	ReservationTime string  `json:"reservation_time"`
+	Price           float32 `json:"price"`
 }
